@@ -1,4 +1,4 @@
-// swift-tools-version:5.8
+// swift-tools-version:6.0
 
 import PackageDescription
 
@@ -19,7 +19,6 @@ let package = Package(
         )
     ],
     dependencies: [
-        .package(url: "https://github.com/apple/swift-docc-plugin", exact: "1.2.0"),
         .package(url: "https://github.com/krzysztofzablocki/Difference", exact: "1.0.2")
     ],
     targets: [
